chains: Reorganises chains to seperate them from the observer
