+++ pathspider/chains/base.py
@@ -1,0 +1,5 @@
+
+class Chain:
+    def new_flow(self, rec, ip):
+        raise NotImplementedError("Cannot create a new flow with an abstract Observer chain")
+
